profeta-forecaster/requirements.txt
# API Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Forecasting (versões mais recentes para Python 3.14)
prophet>=1.1.5
pandas>=2.2.0
numpy>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Database
supabase>=2.4.0

# Validation
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Utils
python-dotenv>=1.0.1
python-multipart>=0.0.9

# Logging
loguru>=0.7.2
